+++ src/Toolbox.DataAccess/DataAccess.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace Toolbox.DataAccess
+{
+
+    /// <summary>
+    /// The sole purpose of this class is to be used as type for the generic ILogger<typeparamref name="T"/>
+    /// </summary>
+    public class DataAccess
+    {
+    }
+}